+++ src/utils.rs
@@ -1,6 +1,4 @@
-use std::collections::HashSet;
 use std::fs;
-use std::hash::Hash;
 use std::io::{BufRead, Write};
 use std::path::PathBuf;
 use log::{debug, error};
@@ -250,7 +248,7 @@ pub(crate) fn get_input_json_content(input: &ConfigInput, url_str: &String, pers
     }
 }
 
-pub(crate) fn dedup<T: Eq + Hash + Copy>(v: &mut Vec<T>) {
-    let mut uniques = HashSet::new();
-    v.retain(|e| uniques.insert(*e));
-}
+// pub(crate) fn dedup<T: Eq + Hash + Copy>(v: &mut Vec<T>) {
+//     let mut uniques = HashSet::new();
+//     v.retain(|e| uniques.insert(*e));
+// }