+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added Target watch for groups
 * Fixed TLS problem with docker scratch
 * Added simple stats
+* Target Output is now a list of multiple output formats, !breaking change!
 
 Changes in `config.yml`
 ```yaml